# storage/docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "Propify API",
        "version": "1.0.0"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/units": {
            "get": {
                "tags": [
                    "Unit"
                ],
                "summary": "Get a listing of the Units.",
                "description": "Get all Units",
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                },
                                "data": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/definitions/Unit"
                                    }
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Unit"
                ],
                "summary": "Store a newly created Unit in storage",
                "description": "Store Unit",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "description": "Unit that should be stored",
                        "required": false,
                        "schema": {
                            "$ref": "#/definitions/Unit"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                },
                                "data": {
                                    "$ref": "#/definitions/Unit"
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            }
        },
        "/units/{id}": {
            "get": {
                "tags": [
                    "Unit"
                ],
                "summary": "Display the specified Unit",
                "description": "Get Unit",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of Unit",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                },
                                "data": {
                                    "$ref": "#/definitions/Unit"
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Unit"
                ],
                "summary": "Update the specified Unit in storage",
                "description": "Update Unit",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of Unit",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "body",
                        "in": "body",
                        "description": "Unit that should be updated",
                        "required": false,
                        "schema": {
                            "$ref": "#/definitions/Unit"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                },
                                "data": {
                                    "$ref": "#/definitions/Unit"
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Unit"
                ],
                "summary": "Remove the specified Unit from storage",
                "description": "Delete Unit",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of Unit",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                },
                                "data": {
                                    "type": "string"
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "Unit": {
            "required": [
                "name",
                "floor"
            ],
            "properties": {
                "id": {
                    "description": "id",
                    "type": "integer",
                    "format": "int32"
                },
                "building_id": {
                    "description": "building_id",
                    "type": "integer",
                    "format": "int32"
                },
                "type": {
                    "description": "type",
                    "type": "integer",
                    "format": "int32"
                },
                "name": {
                    "description": "name",
                    "type": "string"
                },
                "description": {
                    "description": "description",
                    "type": "string"
                },
                "floor": {
                    "description": "floor",
                    "type": "integer",
                    "format": "int32"
                },
                "monthly_rent": {
                    "description": "monthly_rent",
                    "type": "number",
                    "format": "float"
                },
                "room_no": {
                    "description": "room_no",
                    "type": "string",
                    "format": "string"
                },
                "basement": {
                    "description": "basement",
                    "type": "integer",
                    "format": "int32"
                },
                "attic": {
                    "description": "attic",
                    "type": "integer",
                    "format": "int32"
                },
                "created_at": {
                    "description": "created_at",
                    "type": "string",
                    "format": "date-time"
                },
                "updated_at": {
                    "description": "updated_at",
                    "type": "string",
                    "format": "date-time"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "passport-swaggervel_auth": {
            "type": "oauth2",
            "description": "OAuth2 grant provided by Laravel Passport",
            "flow": "accessCode",
            "authorizationUrl": "/oauth/authorize",
            "tokenUrl": "/oauth/token",
            "scopes": {}
        }
    }
}